pyproject.toml
[project]
name = "whisper-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-whisper>=20240930",
    "tqdm>=4.67.1",
]
